requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
openai>=1.15.0
redis>=5.0.4
gspread==6.1.0
google-auth==2.29.0
httpx==0.27.0
twilio>=8.12.0
python-multipart>=0.0.7